requirements.txt
aiohttp==3.5.3
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
flickrapi==2.4.0
idna==2.8
idna-ssl==1.1.0
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
multidict==4.5.2
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.1.0
requests-toolbelt==0.8.0
simplejson==3.16.0
six==1.12.0
typing-extensions==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
yarl==1.3.0
